chore: change dirname from tests to test-e2e
