+++ frontend/plugins/i18n-head.ts
@@ -1,6 +1,8 @@
+import type { Router } from "#vue-router";
+
 // Injects the lang attribute in all Nuxt <Head></Head> components.
 export default defineNuxtPlugin((nuxtApp) => {
-  const router = nuxtApp.vueApp.config.globalProperties.$router;
+  const router: Router = nuxtApp.vueApp.config.globalProperties.$router;
   router.afterEach((to) => {
     const languageCode = to.fullPath.split("/")[1] || "en";
     document.documentElement.setAttribute("lang", languageCode);
