+++ frontend/test-e2e/component-objects/OrganizationMenu.ts
@@ -3,15 +3,19 @@ import { PageObjectBase } from "../utils/PageObjectBase";
 
 const locators = {
   TOGGLE: "#submenu",
-  ABOUT: "#event-about",
-  TEAM: "#event-team",
-  RESOURCES: "#event-resources",
-  TASKS: "#event-tasks",
-  DISCUSSION: "#event-discussion",
-  SETTINGS: "#event-settings",
+  OPTIONS: "#submenu ul",
+  ABOUT: "#org-about",
+  EVENTS: "#org-events",
+  GROUPS: "#org-groups",
+  RESOURCES: "#org-resources",
+  FAQ: "#org-faq",
+  AFFILIATES: "#org-affiliates",
+  TASKS: "#org-tasks",
+  DISCUSSIONS: "#org-discussions",
+  SETTINGS: "#org-settings",
 };
 
-export class EventMenu extends PageObjectBase {
+export class OrganizationMenu extends PageObjectBase {
   constructor(page: Page) {
     super(page, locators);
   }
@@ -20,24 +24,40 @@ export class EventMenu extends PageObjectBase {
     return this.getLocator("TOGGLE");
   }
 
+  get options(): Locator {
+    return this.getLocator("OPTIONS");
+  }
+
   get about(): Locator {
     return this.getLocator("ABOUT");
   }
 
-  get team(): Locator {
-    return this.getLocator("TEAM");
+  get events(): Locator {
+    return this.getLocator("EVENTS");
+  }
+
+  get groups(): Locator {
+    return this.getLocator("GROUPS");
   }
 
   get resources(): Locator {
     return this.getLocator("RESOURCES");
   }
 
+  get faq(): Locator {
+    return this.getLocator("FAQ");
+  }
+
+  get affiliates(): Locator {
+    return this.getLocator("AFFILIATES");
+  }
+
   get tasks(): Locator {
     return this.getLocator("TASKS");
   }
 
-  get discussion(): Locator {
-    return this.getLocator("DISCUSSION");
+  get discussions(): Locator {
+    return this.getLocator("DISCUSSIONS");
   }
 
   get settings(): Locator {
@@ -50,7 +70,7 @@ export class EventMenu extends PageObjectBase {
     }
   }
 
-  async close(): Promise<void> {
+  override async close(): Promise<void> {
     if (await this.isOpen()) {
       await this.toggle.click();
     }
@@ -60,7 +80,7 @@ export class EventMenu extends PageObjectBase {
     return (await this.toggle.getAttribute("aria-expanded")) === "true";
   }
 
-  async isVisible(): Promise<boolean> {
+  override async isVisible(): Promise<boolean> {
     return await this.toggle.isVisible();
   }
 
@@ -69,13 +89,11 @@ export class EventMenu extends PageObjectBase {
   }
 
   async getActiveSelectedOption(): Promise<string> {
+    // if mobile, then "[data-headlessui-state='active selected']"
+    // if desktop, then class includes style-menu-option-cta
     const selector = (await this.isMobile())
       ? "[data-headlessui-state='active selected']"
       : ".style-menu-option-cta";
     return (await this.options.locator(selector).textContent()) || "";
-  }
-
-  async selectOption(option: Locator): Promise<void> {
-    await option.click();
   }
 }
