+++ frontend/test-e2e/component-objects/SearchBar.ts
